Move test store into a shared utils file
